+++ moban/engine_handlebars.py
@@ -9,13 +9,13 @@ from lml.loader import scan_plugins_regex
 from lml.plugin import PluginInfo
 from moban.base_engine import BaseEngine
 from moban.engine_factory import (
-    MOBAN_ALL,
-    BUILTIN_EXENSIONS,
     Context,
     expand_template_directory,
     expand_template_directories,
-    verify_the_existence_of_directories
+    verify_the_existence_of_directories,
+    Strategy,
 )
+from moban.engine import MOBAN_ALL, BUILTIN_EXENSIONS
 
 from pybars import Compiler
 
@@ -39,11 +39,28 @@ class EngineHandlebars(BaseEngine):
                 return os.path.abspath(os.path.join(directory, template_file))
         raise exceptions.FileNotFound(template_file)
 
+    def render_to_files(self, array_of_param_tuple):
+        sta = Strategy(array_of_param_tuple)
+        sta.process()
+        choice = sta.what_to_do()
+        if choice == Strategy.DATA_FIRST:
+            self._render_with_finding_data_first(sta.data_file_index)
+        else:
+            self._render_with_finding_template_first(sta.template_file_index)
+
+    def _file_permissions_copy(self, template_file, output_file):
+        true_template_file = template_file
+        for a_template_dir in self.template_dirs:
+            true_template_file = os.path.join(a_template_dir, template_file)
+            if os.path.exists(true_template_file):
+                break
+        utils.file_permissions_copy(true_template_file, output_file)
+
     def render_to_file(self, template_file, data_file, output_file):
         template_file = self.find_template_file(template_file)
         with open(template_file, "r") as source:
             if sys.version_info[0] < 3:
-                template = Compiler().compile(unicode(source.read())) # noqa
+                template = Compiler().compile(unicode(source.read()))  # noqa
             else:
                 template = Compiler().compile(source.read())
         data = self.context.get_data(data_file)
@@ -77,7 +94,7 @@ class EngineHandlebars(BaseEngine):
         template_file = self.find_template_file(template)
         with open(template_file, "r") as source:
             if sys.version_info[0] < 3:
-                template = Compiler().compile(unicode(source.read())) # noqa
+                template = Compiler().compile(unicode(source.read()))  # noqa
             else:
                 template = Compiler().compile(source.read())
         rendered_content = ''.join(template(data))
